+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - [OpenAPI] Add Blueprinter association parsing with cycle detection (#291)
 - [OpenAPI] Add support for the `view:` option in Blueprinter parser.
 - Add `Rage::Daemon` (#339).
+- Enable non-blocking process monitoring (#341).
 
 ### Fixed